+++ classes.py
@@ -1,6 +1,3 @@
-import json
-
-
 class Category:
     """Class representing a category of products."""
     total_categories = 0
@@ -10,7 +7,7 @@ class Category:
         """Initialize a Category object with name and description."""
         self.name = name
         self.description = description
-        self.products = []
+        self._products = []  # Приватный атрибут для хранения товаров
         Category.total_categories += 1
 
     @classmethod
@@ -19,8 +16,16 @@ class Category:
         _category = cls(data['name'], data['description'])
         for product_data in data.get('products', []):
             product = Product.from_dict(product_data)
-            _category.products.append(product)
+            _category.add_product(product)  # Используем метод добавления продукта
         return _category
+
+    def add_product(self, product):
+        """Add a product to the category."""
+        self._products.append(product)  # Добавляем продукт в список товаров категории
+
+    def get_products(self):
+        """Get the list of products in the category."""
+        return self._products
 
 
 class Product:
@@ -37,25 +42,3 @@ class Product:
     def from_dict(cls, data):
         """Create a Product object from a dictionary."""
         return cls(data['name'], data['description'], data['price'], data['quantity'])
-
-
-def load_data_from_json(file_path):
-    """Load data from a JSON file and create Category objects."""
-    categories = []
-
-    with open(file_path, 'r', encoding='utf-8') as file:
-        data = json.load(file)
-
-    for category_data in data:
-        category = Category.from_dict(category_data)
-        categories.append(category)
-
-    return categories
-
-
-loaded_categories = load_data_from_json('products.json')
-for category in loaded_categories:
-    print(f"Category: {category.name}, Description: {category.description}")
-    print("Products:")
-    for product in category.products:
-        print(f"- {product.name}: {product.description}, Price: {product.price}, Quantity: {product.quantity}")
